requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.22
SQLAlchemy==2.0.36
psycopg2-binary==2.9.9

python-dotenv==1.0.1
pydantic==2.9.2

python-jose==3.3.0
passlib[bcrypt]==1.7.4

alembic==1.14.0

redis==5.1.0

httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0

pydantic-settings==2.6.1
